# src-tauri/openapi.json
{
  "components": {
    "schemas": {
...
      "ClientDisconnect": {
        "properties": {
          "clientId": {
            "type": "string"
          }
        },
        "required": [
          "clientId"
        ],
        "type": "object"
      },
...
      "EventConfig": {
        "description": "事件配置",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "defaultMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "handler": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "pollingEnabled": {
            "type": "boolean"
          },
          "pollingInterval": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "serverId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "serverId",
          "name"
        ],
        "type": "object"
      },
      "EventId": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "EventStatus": {
        "description": "事件运行状态",
        "enum": [
          "enabled",
          "disabled"
        ],
        "type": "string"
      },
      "EventToggle": {
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "type": "object"
      },
...
      "SendBody": {
        "properties": {
          "clientId": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {},
          "event": {
            "type": "string"
          },
          "messageType": {
            "type": [
              "string",
              "null"
            ]
          },
          "serverId": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetId": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetType": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "event"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Socket 服务管理平台后端 REST API（OpenAPI 3.1）。鉴权：Bearer token（`SSM_AUTH_ENABLED=1` 时生效）；`GET /api/auth/bootstrap` 回环自举获取 admin token。",
    "license": {
      "name": ""
    },
    "title": "NexSocket Studio · 管理面 API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/client/disconnect": {
      "post": {
        "operationId": "client_disconnect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDisconnect"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/client/send": {
      "post": {
        "operationId": "client_send",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/clients": {
      "get": {
        "operationId": "get_clients",
        "parameters": [
          {
            "description": "按服务 ID 过滤",
            "in": "query",
            "name": "serverId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/events": {
      "get": {
        "operationId": "get_events",
        "parameters": [
          {
            "description": "按服务 ID 过滤",
            "in": "query",
            "name": "serverId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/events/add": {
      "post": {
        "operationId": "event_add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/events/remove": {
      "post": {
        "operationId": "event_remove",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/events/toggle": {
      "post": {
        "operationId": "event_toggle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/events/update": {
      "post": {
        "operationId": "event_update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
    "/api/send-message": {
      "post": {
        "operationId": "send_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
  },
  "servers": [
    {
      "description": "本地回环（默认绑定）",
      "url": "http://127.0.0.1:3080"
    }
  ]
}